Move ops directly into ops.rs
This makes sense to me because we don't plan on implementing operations
for affine points, so removing the level of nesting is a good change.
